myenv/imdbproject/requirement.txt
Django==1.8.3
argparse==1.2.1
distribute==0.7.3
django-allauth==0.21.0
django-crispy-forms==1.4.0
django-datatables-view==1.12.1
django-filter==0.10.0
django-multiselectfield==0.1.3
django-select-multiple-field==0.4.0
oauthlib==0.7.2
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
